Rename (project-)update to -step to avoid confusion
The term  update could easily be confused with updating data or so.
Another option considered was "entry" but that could also be confused
with an entry in the database. Step is not ideal, because the project
updates/entries/steps are not necessarily the steps of a project,
but could also be just notes or so and user-facing we're still talking
about updates, but the confusion was real and I didn't come up with
a better term.